requirements.txt
streamlit
langchain-google-genai
langchain===1.2.6
langchain-community
faiss-cpu
pypdf
youtube-transcript-api
beautifulsoup4
python-dotenv
langchain-text-splitters